% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algoritmoGenetico.R
\name{algoritmo_genetico}
\alias{algoritmo_genetico}
\title{Implementación de un Algoritmo Genético para el problema TSP}
\usage{
algoritmo_genetico(
  matrizDistancias,
  tamPoblacion,
  numGeneraciones,
  porcentajeElistismo
)
}
\arguments{
\item{matrizDistancias}{Matriz de distancias}

\item{tamPoblacion}{Tamaño de la población.}

\item{numGeneraciones}{Número de generaciones a realizar.}

\item{porcentajeElistismo}{Porcentaje de la población que se conserva en cada generación.}
}
\value{
Lista con los resultados del algoritmo genético.
}
\description{
Implementación de un Algoritmo Genético para el problema TSP
}
