% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/busquedaTabu.R
\name{busqueda_tabu}
\alias{busqueda_tabu}
\title{Búsqueda Tabú para resolver el problema del viajante de comercio (TSP)}
\usage{
busqueda_tabu(
  matrizDistancias,
  maxIteraciones,
  permanenciaListaTabu,
  maxIterSinMejora
)
}
\arguments{
\item{matrizDistancias}{matriz de distancias}

\item{maxIteraciones}{Número máximo de iteraciones.}

\item{permanenciaListaTabu}{Número de iteraciones que un movimiento permanece en la lista tabú.}

\item{maxIterSinMejora}{Número máximo de iteraciones sin mejora.}
}
\value{
Una lista con la mejor solución encontrada, su coste, el tiempo que tardo y la última lista tabú.
}
\description{
Implementa la búsqueda tabú, basada en la matriz de distancias.
}
