% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aco.R
\name{optimizacion_colonia_hormigas}
\alias{optimizacion_colonia_hormigas}
\title{Implementación de la metaheurística de colonia de hormigas para el TSP}
\usage{
optimizacion_colonia_hormigas(
  matrizDistancias,
  numHormigas,
  maxIteraciones,
  tasaEvaporacion,
  alfa,
  beta
)
}
\arguments{
\item{matrizDistancias}{Matriz de distancias.}

\item{numHormigas}{Número de hormigas a utilizar.}

\item{maxIteraciones}{Número máximo de iteraciones.}

\item{tasaEvaporacion}{Tasa de evaporación de las feromonas.}

\item{alfa}{Parámetro alfa (importancia de las feromonas).}

\item{beta}{Parámetro beta (importancia de las distancias).}
}
\value{
Lista con la mejor solución encontrada, su coste y métricas del proceso.
}
\description{
Implementación de la metaheurística de colonia de hormigas para el TSP
}
