% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templadoSimulado.R
\name{templado_simulado}
\alias{templado_simulado}
\title{Templado Simulado para resolver el problema del viajante de comercio (TSP)}
\usage{
templado_simulado(
  matrizDistancias,
  temperaturaInicial,
  tasaEnfriamiento,
  temperaturaParada,
  maxIteracionesSinCambios,
  maxIteraciones
)
}
\arguments{
\item{matrizDistancias}{Matriz de distancias}

\item{temperaturaInicial}{Temperatura inicial del templado simulado.}

\item{tasaEnfriamiento}{La tasa de enfriamiento de la temperatura (valor entre 0 y 1).}

\item{temperaturaParada}{Temperatura mínima para detener el algoritmo.}

\item{maxIteracionesSinCambios}{Máximo número de iteraciones sin cambios permitidas.}

\item{maxIteraciones}{Número máximo de iteraciones por temperatura.}
}
\value{
Una lista con la mejor solución, su coste, y datos del proceso (temperaturas, costes, probabilidad de aceptacion).
}
\description{
Este algoritmo utiliza el metodo de templado simualdo para encontrar una solución aproximada al TSP.
}
